Break image_generators.py into a submodule with each generator type in a different file.
