+++ deepcell/image_generators/sample.py
@@ -1,0 +1,680 @@
+# Copyright 2016-2019 The Van Valen Lab at the California Institute of
+# Technology (Caltech), with support from the Paul Allen Family Foundation,
+# Google, & National Institutes of Health (NIH) under Grant U24CA224309-01.
+# All rights reserved.
+#
+# Licensed under a modified Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.github.com/vanvalenlab/deepcell-tf/LICENSE
+#
+# The Work provided may be used for non-commercial academic purposes only.
+# For any other use of the Work, including commercial use, please contact:
+# vanvalenlab@gmail.com
+#
+# Neither the name of Caltech nor the names of its contributors may be used
+# to endorse or promote products derived from this software without specific
+# prior written permission.
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Sample based data generators."""
+
+from __future__ import absolute_import
+from __future__ import print_function
+from __future__ import division
+
+import os
+
+import numpy as np
+
+from tensorflow.python.keras import backend as K
+from tensorflow.python.keras.utils import to_categorical
+from tensorflow.python.keras.utils import conv_utils
+from tensorflow.python.keras.preprocessing.image import array_to_img
+
+from deepcell.image_generators import _transform_masks
+from deepcell.image_generators import Iterator
+from deepcell.image_generators import ImageDataGenerator
+
+from deepcell.image_generators import MovieDataGenerator
+
+from deepcell.utils.data_utils import sample_label_movie
+from deepcell.utils.data_utils import sample_label_matrix
+
+
+class ImageSampleArrayIterator(Iterator):
+    """Iterator yielding data from a sampled Numpy array.
+    Sampling will generate a window_size image classifying the center pixel,
+
+    Args:
+        train_dict: dictionary consisting of numpy arrays for X and y.
+        image_data_generator: Instance of ImageDataGenerator
+            to use for random transformations and normalization.
+        batch_size: Integer, size of a batch.
+        shuffle: Boolean, whether to shuffle the data between epochs.
+        window_size: size of sampling window around each pixel
+        balance_classes: balance class representation when sampling
+        max_class_samples: maximum number of samples per class.
+        seed: Random seed for data shuffling.
+        data_format: String, one of 'channels_first', 'channels_last'.
+        save_to_dir: Optional directory where to save the pictures
+            being yielded, in a viewable format. This is useful
+            for visualizing the random transformations being
+            applied, for debugging purposes.
+        save_prefix: String prefix to use for saving sample
+            images (if save_to_dir is set).
+        save_format: Format to use for saving sample images
+            (if save_to_dir is set).
+    """
+
+    def __init__(self,
+                 train_dict,
+                 image_data_generator,
+                 batch_size=32,
+                 shuffle=False,
+                 window_size=(30, 30),
+                 transform=None,
+                 transform_kwargs={},
+                 balance_classes=False,
+                 max_class_samples=None,
+                 seed=None,
+                 data_format='channels_last',
+                 save_to_dir=None,
+                 save_prefix='',
+                 save_format='png'):
+        X, y = train_dict['X'], train_dict['y']
+        if y is not None and X.shape[0] != y.shape[0]:
+            raise ValueError('Training batches and labels should have the same'
+                             'length. Found X.shape: {} y.shape: {}'.format(
+                                 X.shape, y.shape))
+        self.x = np.asarray(X, dtype=K.floatx())
+
+        if self.x.ndim != 4:
+            raise ValueError('Input data in `ImageSampleArrayIterator` '
+                             'should have rank 4. You passed an array '
+                             'with shape', self.x.shape)
+
+        window_size = conv_utils.normalize_tuple(window_size, 2, 'window_size')
+
+        y = _transform_masks(y, transform, data_format=data_format, **transform_kwargs)
+
+        pixels_x, pixels_y, batch, y = sample_label_matrix(
+            y=y,
+            padding='valid',
+            window_size=window_size,
+            max_training_examples=None,
+            data_format=data_format)
+
+        self.y = y
+        self.channel_axis = 3 if data_format == 'channels_last' else 1
+        self.batch = batch
+        self.pixels_x = pixels_x
+        self.pixels_y = pixels_y
+        self.win_x = window_size[0]
+        self.win_y = window_size[1]
+        self.image_data_generator = image_data_generator
+        self.data_format = data_format
+        self.save_to_dir = save_to_dir
+        self.save_prefix = save_prefix
+        self.save_format = save_format
+
+        self.class_balance(max_class_samples, balance_classes, seed=seed)
+
+        self.y = to_categorical(self.y).astype('int32')
+        super(ImageSampleArrayIterator, self).__init__(
+            len(self.y), batch_size, shuffle, seed)
+
+    def _sample_image(self, b, px, py):
+        wx = self.win_x
+        wy = self.win_y
+
+        if self.channel_axis == 1:
+            sampled = self.x[b, :, px - wx:px + wx + 1, py - wy:py + wy + 1]
+        else:
+            sampled = self.x[b, px - wx:px + wx + 1, py - wy:py + wy + 1, :]
+
+        return sampled
+
+    def class_balance(self, max_class_samples=None, downsample=False, seed=None):
+        """Balance classes based on the number of samples of each class.
+
+        Args:
+            max_class_samples: if not None, a maximum count for each class
+            downsample: if True, all sample sizes will be the rarest count
+            seed: random state initalization
+
+        Returns:
+            Does not return anything but shuffles and resizes the sample size
+        """
+        balanced_indices = []
+
+        unique_b = np.unique(self.batch)
+
+        if max_class_samples is not None:
+            max_class_samples = int(max_class_samples // len(unique_b))
+
+        for b in unique_b:
+            batch_y = self.y[self.batch == b]
+            unique, counts = np.unique(batch_y, return_counts=True)
+            min_index = np.argmin(counts)
+            n_samples = counts[min_index]
+
+            if max_class_samples is not None and max_class_samples < n_samples:
+                n_samples = max_class_samples
+
+            for class_label in unique:
+                non_rand_ind = ((self.batch == b) & (self.y == class_label)).nonzero()[0]
+
+                if downsample:
+                    size = n_samples
+                elif max_class_samples:
+                    size = min(max_class_samples, len(non_rand_ind))
+                else:
+                    size = len(non_rand_ind)
+
+                index = np.random.choice(non_rand_ind, size=size, replace=False)
+                balanced_indices.extend(index)
+
+        np.random.seed(seed=seed)
+        np.random.shuffle(balanced_indices)
+
+        self.batch = self.batch[balanced_indices]
+        self.pixels_x = self.pixels_x[balanced_indices]
+        self.pixels_y = self.pixels_y[balanced_indices]
+        self.y = self.y[balanced_indices]
+
+    def _get_batches_of_transformed_samples(self, index_array):
+        if self.channel_axis == 1:
+            batch_x = np.zeros((len(index_array),
+                                self.x.shape[self.channel_axis],
+                                2 * self.win_x + 1,
+                                2 * self.win_y + 1))
+        else:
+            batch_x = np.zeros((len(index_array),
+                                2 * self.win_x + 1,
+                                2 * self.win_y + 1,
+                                self.x.shape[self.channel_axis]))
+
+        for i, j in enumerate(index_array):
+            b, px, py = self.batch[j], self.pixels_x[j], self.pixels_y[j]
+            x = self._sample_image(b, px, py)
+            x = self.image_data_generator.random_transform(x.astype(K.floatx()))
+            x = self.image_data_generator.standardize(x)
+
+            batch_x[i] = x
+
+        if self.save_to_dir:
+            for i, j in enumerate(index_array):
+                if self.data_format == 'channels_first':
+                    img_x = np.expand_dims(batch_x[i, 0, ...], 0)
+                else:
+                    img_x = np.expand_dims(batch_x[i, ..., 0], -1)
+                img = array_to_img(img_x, self.data_format, scale=True)
+                fname = '{prefix}_{index}_{hash}.{format}'.format(
+                    prefix=self.save_prefix,
+                    index=j,
+                    hash=np.random.randint(1e4),
+                    format=self.save_format)
+                img.save(os.path.join(self.save_to_dir, fname))
+
+        if self.y is None:
+            return batch_x
+        batch_y = self.y[index_array]
+        return batch_x, batch_y
+
+    def next(self):
+        """For python 2.x. Returns the next batch.
+        """
+        # Keeps under lock only the mechanism which advances
+        # the indexing of each batch.
+        with self.lock:
+            index_array = next(self.index_generator)
+        # The transformation of images is not under thread lock
+        # so it can be done in parallel
+        return self._get_batches_of_transformed_samples(index_array)
+
+
+class SampleDataGenerator(ImageDataGenerator):
+    """Generates batches of tensor image data with real-time data augmentation.
+
+    The data will be looped over (in batches).
+
+    Args:
+        featurewise_center: boolean, set input mean to 0 over the dataset,
+            feature-wise.
+        samplewise_center: boolean, set each sample mean to 0.
+        featurewise_std_normalization: boolean, divide inputs by std
+            of the dataset, feature-wise.
+        samplewise_std_normalization: boolean, divide each input by its std.
+        zca_epsilon: epsilon for ZCA whitening. Default is 1e-6.
+        zca_whitening: boolean, apply ZCA whitening.
+        rotation_range: int, degree range for random rotations.
+        width_shift_range: float, 1-D array-like or int
+
+            - float: fraction of total width, if < 1, or pixels if >= 1.
+            - 1-D array-like: random elements from the array.
+            - int: integer number of pixels from interval
+              (-width_shift_range, +width_shift_range)
+            - With width_shift_range=2 possible values are ints [-1, 0, +1],
+              same as with width_shift_range=[-1, 0, +1],
+              while with width_shift_range=1.0 possible values are floats in
+              the interval [-1.0, +1.0).
+
+        shear_range: float, shear Intensity
+            (Shear angle in counter-clockwise direction in degrees)
+        zoom_range: float or [lower, upper], Range for random zoom.
+            If a float, [lower, upper] = [1-zoom_range, 1+zoom_range].
+        channel_shift_range: float, range for random channel shifts.
+        fill_mode: One of {"constant", "nearest", "reflect" or "wrap"}.
+
+            Default is 'nearest'. Points outside the boundaries of the input
+            are filled according to the given mode:
+
+                - 'constant': kkkkkkkk|abcd|kkkkkkkk (cval=k)
+                - 'nearest':  aaaaaaaa|abcd|dddddddd
+                - 'reflect':  abcddcba|abcd|dcbaabcd
+                - 'wrap':  abcdabcd|abcd|abcdabcd
+
+        cval: float or int, value used for points outside the boundaries
+            when fill_mode = "constant".
+        horizontal_flip: boolean, randomly flip inputs horizontally.
+        vertical_flip: boolean, randomly flip inputs vertically.
+        rescale: rescaling factor. Defaults to None. If None or 0, no rescaling
+            is applied, otherwise we multiply the data by the value provided
+            (before applying any other transformation).
+        preprocessing_function: function that will be implied on each input.
+            The function will run after the image is resized and augmented.
+            The function should take one argument:
+            one image (Numpy tensor with rank 3),
+            and should output a Numpy tensor with the same shape.
+
+        data_format: One of {"channels_first", "channels_last"}.
+
+            - "channels_last" mode means that the images should have shape
+              (samples, height, width, channels),
+            - "channels_first" mode means that the images should have shape
+              (samples, channels, height, width).
+            - It defaults to the image_data_format value found in your
+              Keras config file at "~/.keras/keras.json".
+            - If you never set it, then it will be "channels_last".
+
+        validation_split: float, fraction of images reserved for validation
+            (strictly between 0 and 1).
+    """
+
+    def flow(self,
+             train_dict,
+             batch_size=32,
+             shuffle=True,
+             transform=None,
+             transform_kwargs={},
+             window_size=(30, 30),
+             balance_classes=False,
+             max_class_samples=None,
+             seed=None,
+             save_to_dir=None,
+             save_prefix='',
+             save_format='png'):
+        """Generates batches of augmented/normalized data with given arrays.
+
+        Args:
+            train_dict: dictionary consisting of numpy arrays for X and y.
+            image_data_generator: Instance of ImageDataGenerator
+                to use for random transformations and normalization.
+            batch_size: Integer, size of a batch.
+            shuffle: Boolean, whether to shuffle the data between epochs.
+            window_size: size of sampling window around each pixel
+            balance_classes: balance class representation when sampling
+            max_class_samples: maximum number of samples per class.
+            seed: Random seed for data shuffling.
+            data_format: String, one of 'channels_first', 'channels_last'.
+            save_to_dir: Optional directory where to save the pictures
+                being yielded, in a viewable format. This is useful
+                for visualizing the random transformations being
+                applied, for debugging purposes.
+            save_prefix: String prefix to use for saving sample
+                images (if save_to_dir is set).
+            save_format: Format to use for saving sample images
+                (if save_to_dir is set).
+        """
+        return ImageSampleArrayIterator(
+            train_dict,
+            self,
+            batch_size=batch_size,
+            shuffle=shuffle,
+            transform=transform,
+            transform_kwargs=transform_kwargs,
+            window_size=window_size,
+            balance_classes=balance_classes,
+            max_class_samples=max_class_samples,
+            seed=seed,
+            data_format=self.data_format,
+            save_to_dir=save_to_dir,
+            save_prefix=save_prefix,
+            save_format=save_format)
+
+
+class SampleMovieArrayIterator(Iterator):
+    """Iterator yielding data from two 5D Numpy arrays (X and y).
+
+    Sampling will generate a window_size voxel classifying the center pixel,
+
+    Args:
+        train_dict: dictionary consisting of numpy arrays for X and y.
+        movie_data_generator: Instance of MovieDataGenerator
+            to use for random transformations and normalization.
+        batch_size: Integer, size of a batch.
+        shuffle: Boolean, whether to shuffle the data between epochs.
+        window_size: size of sampling window around each pixel
+        balance_classes: balance class representation when sampling
+        max_class_samples: maximum number of samples per class.
+        seed: Random seed for data shuffling.
+        data_format: String, one of 'channels_first', 'channels_last'.
+        save_to_dir: Optional directory where to save the pictures
+            being yielded, in a viewable format. This is useful
+            for visualizing the random transformations being
+            applied, for debugging purposes.
+        save_prefix: String prefix to use for saving sample
+            images (if save_to_dir is set).
+        save_format: Format to use for saving sample images
+            (if save_to_dir is set).
+    """
+
+    def __init__(self,
+                 train_dict,
+                 movie_data_generator,
+                 batch_size=32,
+                 shuffle=False,
+                 transform=None,
+                 transform_kwargs={},
+                 balance_classes=False,
+                 max_class_samples=None,
+                 window_size=(30, 30, 5),
+                 seed=None,
+                 data_format='channels_last',
+                 save_to_dir=None,
+                 save_prefix='',
+                 save_format='png'):
+        X, y = train_dict['X'], train_dict['y']
+        if y is not None and X.shape[0] != y.shape[0]:
+            raise ValueError('`X` (movie data) and `y` (labels) '
+                             'should have the same size. Found '
+                             'Found x.shape = {}, y.shape = {}'.format(
+                                 X.shape, y.shape))
+        self.channel_axis = 4 if data_format == 'channels_last' else 1
+        self.time_axis = 1 if data_format == 'channels_last' else 2
+        self.x = np.asarray(X, dtype=K.floatx())
+        y = _transform_masks(y, transform,
+                             data_format=data_format,
+                             **transform_kwargs)
+
+        if self.x.ndim != 5:
+            raise ValueError('Input data in `SampleMovieArrayIterator` '
+                             'should have rank 5. You passed an array '
+                             'with shape', self.x.shape)
+
+        window_size = conv_utils.normalize_tuple(window_size, 3, 'window_size')
+
+        pixels_z, pixels_x, pixels_y, batch, y = sample_label_movie(
+            y=y,
+            padding='valid',
+            window_size=window_size,
+            max_training_examples=None,
+            data_format=data_format)
+
+        self.y = y
+        self.win_x = window_size[0]
+        self.win_y = window_size[1]
+        self.win_z = window_size[2]
+        self.pixels_x = pixels_x
+        self.pixels_y = pixels_y
+        self.pixels_z = pixels_z
+        self.batch = batch
+        self.movie_data_generator = movie_data_generator
+        self.data_format = data_format
+        self.save_to_dir = save_to_dir
+        self.save_prefix = save_prefix
+        self.save_format = save_format
+
+        self.class_balance(max_class_samples, balance_classes, seed=seed)
+
+        self.y = to_categorical(self.y).astype('int32')
+        super(SampleMovieArrayIterator, self).__init__(
+            len(self.y), batch_size, shuffle, seed)
+
+    def _sample_image(self, b, pz, px, py):
+        wx = self.win_x
+        wy = self.win_y
+        wz = self.win_z
+
+        if self.channel_axis == 1:
+            sampled = self.x[b, :, pz - wz:pz + wz + 1, px - wx:px + wx + 1, py - wy:py + wy + 1]
+        else:
+            sampled = self.x[b, pz - wz:pz + wz + 1, px - wx:px + wx + 1, py - wy:py + wy + 1, :]
+
+        return sampled
+
+    def class_balance(self, max_class_samples=None, downsample=False, seed=None):
+        """Balance classes based on the number of samples of each class
+
+        Args:
+            max_class_samples: if not None, a maximum count for each class
+            downsample: if True, all sample sizes will be the rarest count
+            seed: random state initalization
+        """
+        balanced_indices = []
+
+        unique_b = np.unique(self.batch)
+
+        if max_class_samples is not None:
+            max_class_samples = int(max_class_samples // len(unique_b))
+
+        for b in unique_b:
+            batch_y = self.y[self.batch == b]
+            unique, counts = np.unique(batch_y, return_counts=True)
+            min_index = np.argmin(counts)
+            n_samples = counts[min_index]
+
+            if max_class_samples is not None and max_class_samples < n_samples:
+                n_samples = max_class_samples
+
+            for class_label in unique:
+                non_rand_ind = ((self.batch == b) & (self.y == class_label)).nonzero()[0]
+
+                if downsample:
+                    size = n_samples
+                elif max_class_samples:
+                    size = min(max_class_samples, len(non_rand_ind))
+                else:
+                    size = len(non_rand_ind)
+
+                index = np.random.choice(non_rand_ind, size=size, replace=False)
+                balanced_indices.extend(index)
+
+        np.random.seed(seed=seed)
+        np.random.shuffle(balanced_indices)
+
+        # Save the upsampled results
+        self.batch = self.batch[balanced_indices]
+        self.pixels_z = self.pixels_z[balanced_indices]
+        self.pixels_x = self.pixels_x[balanced_indices]
+        self.pixels_y = self.pixels_y[balanced_indices]
+        self.y = self.y[balanced_indices]
+
+    def _get_batches_of_transformed_samples(self, index_array):
+        if self.channel_axis == 1:
+            batch_x = np.zeros((len(index_array),
+                                self.x.shape[self.channel_axis],
+                                2 * self.win_z + 1,
+                                2 * self.win_x + 1,
+                                2 * self.win_y + 1))
+        else:
+            batch_x = np.zeros((len(index_array),
+                                2 * self.win_z + 1,
+                                2 * self.win_x + 1,
+                                2 * self.win_y + 1,
+                                self.x.shape[self.channel_axis]))
+
+        for i, j in enumerate(index_array):
+            b, pz, px, py = self.batch[j], self.pixels_z[j], self.pixels_x[j], self.pixels_y[j]
+            x = self._sample_image(b, pz, px, py)
+            x = self.movie_data_generator.random_transform(x.astype(K.floatx()))
+            x = self.movie_data_generator.standardize(x)
+
+            batch_x[i] = x
+
+        if self.save_to_dir:
+            time_axis = 2 if self.data_format == 'channels_first' else 1
+            for i, j in enumerate(index_array):
+                for frame in range(batch_x.shape[time_axis]):
+                    if time_axis == 2:
+                        img = batch_x[i, :, frame]
+                    else:
+                        img = batch_x[i, frame]
+                    img = array_to_img(img, self.data_format, scale=True)
+                    fname = '{prefix}_{index}_{hash}.{format}'.format(
+                        prefix=self.save_prefix,
+                        index=j,
+                        hash=np.random.randint(1e4),
+                        format=self.save_format)
+                    img.save(os.path.join(self.save_to_dir, fname))
+
+        if self.y is None:
+            return batch_x
+        batch_y = self.y[index_array]
+        return batch_x, batch_y
+
+    def next(self):
+        """For python 2.x. Returns the next batch.
+        """
+        # Keeps under lock only the mechanism which advances
+        # the indexing of each batch.
+        with self.lock:
+            index_array = next(self.index_generator)
+        # The transformation of images is not under thread lock
+        # so it can be done in parallel
+        return self._get_batches_of_transformed_samples(index_array)
+
+
+class SampleMovieDataGenerator(MovieDataGenerator):
+    """Generates batches of tensor image data with real-time data augmentation.
+
+    The data will be looped over (in batches).
+
+    Args:
+        featurewise_center: boolean, set input mean to 0 over the dataset,
+            feature-wise.
+        samplewise_center: boolean, set each sample mean to 0.
+        featurewise_std_normalization: boolean, divide inputs by std
+            of the dataset, feature-wise.
+        samplewise_std_normalization: boolean, divide each input by its std.
+        zca_epsilon: epsilon for ZCA whitening. Default is 1e-6.
+        zca_whitening: boolean, apply ZCA whitening.
+        rotation_range: int, degree range for random rotations.
+        width_shift_range: float, 1-D array-like or int
+
+            - float: fraction of total width, if < 1, or pixels if >= 1.
+            - 1-D array-like: random elements from the array.
+            - int: integer number of pixels from interval
+              (-width_shift_range, +width_shift_range)
+            - With width_shift_range=2 possible values are ints [-1, 0, +1],
+              same as with width_shift_range=[-1, 0, +1],
+              while with width_shift_range=1.0 possible values are floats in
+              the interval [-1.0, +1.0).
+
+        shear_range: float, shear Intensity
+            (Shear angle in counter-clockwise direction in degrees)
+        zoom_range: float or [lower, upper], Range for random zoom.
+            If a float, [lower, upper] = [1-zoom_range, 1+zoom_range].
+        channel_shift_range: float, range for random channel shifts.
+        fill_mode: One of {"constant", "nearest", "reflect" or "wrap"}.
+
+            Default is 'nearest'. Points outside the boundaries of the input
+            are filled according to the given mode:
+                - 'constant': kkkkkkkk|abcd|kkkkkkkk (cval=k)
+                - 'nearest':  aaaaaaaa|abcd|dddddddd
+                - 'reflect':  abcddcba|abcd|dcbaabcd
+                - 'wrap':  abcdabcd|abcd|abcdabcd
+
+        cval: float or int, value used for points outside the boundaries
+            when fill_mode = "constant".
+        horizontal_flip: boolean, randomly flip inputs horizontally.
+        vertical_flip: boolean, randomly flip inputs vertically.
+        rescale: rescaling factor. Defaults to None. If None or 0, no rescaling
+            is applied, otherwise we multiply the data by the value provided
+            (before applying any other transformation).
+        preprocessing_function: function that will be implied on each input.
+            The function will run after the image is resized and augmented.
+            The function should take one argument:
+            one image (Numpy tensor with rank 3),
+            and should output a Numpy tensor with the same shape.
+        data_format: One of {"channels_first", "channels_last"}.
+
+            - "channels_last" mode means that the images should have shape
+              (samples, height, width, channels),
+            - "channels_first" mode means that the images should have shape
+              (samples, channels, height, width).
+            - It defaults to the image_data_format value found in your
+              Keras config file at "~/.keras/keras.json".
+            - If you never set it, then it will be "channels_last".
+
+        validation_split: float, fraction of images reserved for validation
+            (strictly between 0 and 1).
+    """
+
+    def flow(self,
+             train_dict,
+             batch_size=32,
+             shuffle=True,
+             transform=None,
+             transform_kwargs={},
+             window_size=(30, 30, 5),
+             balance_classes=False,
+             max_class_samples=None,
+             seed=None,
+             save_to_dir=None,
+             save_prefix='',
+             save_format='png'):
+        """Generates batches of augmented/normalized data with given arrays.
+
+        Args:
+            train_dict: dictionary of X and y tensors. Both should be rank 5.
+            window_size: tuple (default: (30, 30 5)).
+                The size of the sampled voxels to generate.
+            batch_size: int (default: 1).
+            shuffle: boolean (default: True).
+            seed: int (default: None).
+            save_to_dir: None or str (default: None).
+                This allows you to optionally specify a directory
+                to which to save the augmented pictures being generated
+                (useful for visualizing what you are doing).
+            save_prefix: str (default: ''). Prefix to use for filenames of
+                saved pictures (only relevant if save_to_dir is set).
+            save_format: one of "png", "jpeg". Default: "png".
+                (only relevant if save_to_dir is set)
+
+        Returns:
+            An Iterator yielding tuples of (x, y) where x is a numpy array
+            of image data and y is a numpy array of labels of the same shape.
+        """
+        return SampleMovieArrayIterator(
+            train_dict,
+            self,
+            batch_size=batch_size,
+            shuffle=shuffle,
+            transform=transform,
+            transform_kwargs=transform_kwargs,
+            window_size=window_size,
+            balance_classes=balance_classes,
+            max_class_samples=max_class_samples,
+            seed=seed,
+            data_format=self.data_format,
+            save_to_dir=save_to_dir,
+            save_prefix=save_prefix,
+            save_format=save_format)
